cloud-movies/requirements.txt
aws-cdk-lib==2.145.0
aws-cdk.aws-lambda-python-alpha>=2.142.1a0
constructs>=10.0.0,<11.0.0
boto3==1.34.122